Removed slack global features
