+++ src/main/java/com/blackducksoftware/integration/hub/alert/channel/slack/SlackChannel.java
@@ -83,18 +83,17 @@ public class SlackChannel extends DistributionChannel<SlackEvent, GlobalSlackCon
     }
 
     @Override
-    public String testMessage() throws IntegrationException {
-        // TODO send test message from global config
+    public String testMessage(final SlackDistributionConfigEntity distributionConfig) throws IntegrationException {
         final String message = "*Test* from _Alert_ application";
-        return sendMessage(message, null);
+        return sendMessage(message, distributionConfig);
     }
 
     private String sendMessage(final String htmlMessage, final SlackDistributionConfigEntity config) throws IntegrationException {
-        final String slackUrl = getGlobalConfigEntity().getWebhook();
+        final String slackUrl = config.getWebhook();
         final ChannelRestConnectionFactory restConnectionFactory = new ChannelRestConnectionFactory(null);
         final RestConnection connection = restConnectionFactory.createUnauthenticatedRestConnection(slackUrl);
         if (connection != null) {
-            final String jsonString = getJsonString(htmlMessage, config.getChannelName(), getGlobalConfigEntity().getChannelUsername());
+            final String jsonString = getJsonString(htmlMessage, config.getChannelName(), config.getChannelUsername());
             final RequestBody body = connection.createJsonRequestBody(jsonString);
 
             final Map<String, String> requestProperties = new HashMap<>();
